requirements.txt
asgiref==3.9.1
Django==5.2.5
djangorestframework==3.16.1
packaging==25.0
pillow==11.3.0
pytesseract==0.3.13
sqlparse==0.5.3
google-generativeai==0.7.2
